change minor things

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ksmorozo <ksmorozo@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2020/10/29 13:03:49 by ksmorozo      #+#    #+#                 */
-/*   Updated: 2020/12/03 00:13:17 by anonymous     ########   odam.nl         */
+/*   Updated: 2020/12/07 13:39:23 by ksmorozo      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int count;
+	size_t count;
 
 	count = 0;
 	while (str[count] != '\0')
